+++ CHANGELOG.md
@@ -2,8 +2,9 @@
 
 ## 0.1.3 (TBA)
 
-- [`<KabelwerkMessage>`](./docs/KabelwerkMessage.md): added the `renderCheckmarks` and `onLongPress` prop.
+- Added the [`<KabelwerkTheme>`](./docs/KabelwerkTheme.md) context.
 - [`<KabelwerkRoom>`](./docs/KabelwerkRoom.md): added the `onReady` prop.
+- [`<KabelwerkMessage>`](./docs/KabelwerkMessage.md): added the `renderCheckmarks` and `onLongPress` prop.
 
 ## 0.1.2 (2022-08-28)